While Running CFD Application, we got the following error
Exception : SMPDeviceImpl.id() invoked by thread " +
                    "outside communicator group


Before Explaining the error and its reason, we explain the way shmdev 
adds threads to its HashMap

 shmdev uses a HashMap variable ids in which it stores the threads using 
method ids.put(thread,rank). In the Init() phase, all threads are put in 
this HashMap and they constitue a communicator group.

Now this code used to run fine for all the user applications which are 
single threaded. But for CFD application, it gave the above descibed 
Exception. We traced the reason for this error. The reason is :: CFD 
application is multithreaded as in the application, a thread is created 
for drawing. 


So we concider the scenario where we try to run CFD application with two 
threads as 

mpjrun.sh -np 2 -dev shmdev CFD

So the ids  HashMap variable will add two threads in HashMap with lets 
say id value 9 and 8 (id can be get using thread.getId()). Now the 
application thread, i-e the thread created within CFD application is not 
registered in this Hashmap, so when this thread tries to call Send() 
Recv(), it gives the above mentioned Exception as this thread never got 
registered. 


What we did to solve this issue was, instead of storing thread in 
HashMap, we stored a thread.getContextClassLoader() in the HashMap. In 
this way, the thread that is created in the CFD application also has the 
same ContextClassLoader as the thread created in the shmdev. So now when 
the application thread tries to call Send() Recv(), it gives no error.

  -- 
Kamran Hameed